add version check to ui
